Car and employee dao using SimpleDataAccessException

+++ CarPark/CarParkPersistence/src/main/java/com/fi/muni/carparkapp/dao/CarDaoImpl.java
@@ -7,6 +7,7 @@
 package com.fi.muni.carparkapp.dao;
 
 import com.fi.muni.carparkapp.entity.Car;
+import com.fi.muni.carparkapp.exceptions.SimpleDataAccessException;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -24,11 +25,18 @@ public class CarDaoImpl implements CarDao{
     
     @Override
     public Car findById(Long id) {
+        if(id <= 0)
+            throw new SimpleDataAccessException("id cannot be < 0");
         return em.find(Car.class, id);
     }
     
     @Override
     public Car findByVin(String vin) {
+        if(vin == null)
+            throw  new SimpleDataAccessException("vin cannot be null");
+        if(vin == "")
+            throw  new SimpleDataAccessException("vin cannot be empty");
+        
         List<Car> resultList = em.createQuery("SELECT c FROM Car c where c.vin = :vin", Car.class)
                 .setParameter("vin", vin).getResultList();
         return resultList.get(0);
@@ -36,11 +44,15 @@ public class CarDaoImpl implements CarDao{
     
     @Override
     public void create(Car c) {
+        if(c == null)
+            throw new SimpleDataAccessException("car cannot be null");
         em.persist(c);
     }
 
     @Override
     public void delete(Car c) {
+        if(c == null)
+            throw new SimpleDataAccessException("car cannot be null");
         em.remove(c);
     }
 
@@ -51,6 +63,8 @@ public class CarDaoImpl implements CarDao{
 
     @Override
     public void update(Car c) {
+        if(c == null)
+            throw new SimpleDataAccessException("car cannot be null");
         em.merge(c);
     }
